feat: Refactor api
